fix: get_column_labels now can return the output as pd.Dataframe
